+++ cnb-builder-shim/internal/devsandbox/processesctl/supervisor_rpc_process_controller.go
@@ -25,8 +25,6 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/pkg/errors"
-
 	"github.com/TencentBlueking/bkpaas/cnb-builder-shim/pkg/appdesc"
 	"github.com/TencentBlueking/bkpaas/cnb-builder-shim/pkg/supervisord/rpc"
 )
@@ -36,7 +34,7 @@ var rpcPort = "9001"
 var rpcAddress = "http://127.0.0.1:9001/RPC2"
 
 var confFilePath = filepath.Join(supervisorDir, "dev.conf")
-
+var controllerType = SupervisorRPC
 var confTmpl = `[unix_http_server]
 file = {{ .RootDir }}/supervisor.sock
 
@@ -61,17 +59,6 @@ redirect_stderr = true
 [inet_http_server]
 port=127.0.0.1:{{ .Port }}
 `
-
-type Process struct {
-	ProcType    string
-	CommandPath string
-}
-
-// ProcessConf is a process config
-type ProcessConf struct {
-	Process
-	ProcLogFile string
-}
 
 // SupervisorConf is a supervisor template conf data
 type SupervisorConf struct {
@@ -125,91 +112,6 @@ func refreshConf(conf *SupervisorConf) error {
 	return tmpl.Execute(file, *conf)
 }
 
-// validateEnvironment validates the environment variables for supervisor conf.
-//
-// see detail environment conf in http://supervisord.org/configuration.html
-// char " and % in environment value will cause supervisord to fail
-func validateEnvironment(procEnvs []appdesc.Env) error {
-	invalidChars := `"%`
-	invalidEnvNames := []string{}
-	for _, env := range procEnvs {
-		if strings.ContainsAny(env.Value, invalidChars) {
-			invalidEnvNames = append(invalidEnvNames, env.Name)
-		}
-	}
-	if len(invalidEnvNames) == 0 {
-		return nil
-	}
-
-	return fmt.Errorf(
-		"environment variables: %s has invalid characters (%s)",
-		strings.Join(invalidEnvNames, ", "),
-		invalidChars,
-	)
-}
-
-// ProcessCtl 用于进程控制的接口
-type ProcessCtl interface {
-	// Status 获取所有进程的状态
-	Status() ([]rpc.ProcessInfo, error)
-	// Start 启动进程（只能操作已存在的进程）
-	Start(name string) error
-	// Stop 停止(不是删除)进程
-	Stop(name string) error
-	// Reload 更新和重启进程列表
-	Reload() error
-	// StopAllProcesses 停止所有进程
-	StopAllProcesses() error
-}
-
-// ControllerType 定义控制器类型
-type ControllerType string
-
-const (
-	RPC ControllerType = "rpc"
-)
-
-// NewProcessController 根据类型返回不同的 ProcessCtl 实现
-func NewProcessController(controllerType ControllerType) (ProcessCtl, error) {
-	switch controllerType {
-	case RPC:
-		return newRPCProcessController()
-	default:
-		return nil, errors.New("unsupported controller type")
-	}
-}
-
-// RPCProcessController ...
-type RPCProcessController struct {
-	client *rpc.Client
-}
-
-// 创建 RPC 类型的 ProcessController
-func newRPCProcessController() (*RPCProcessController, error) {
-	client, err := rpc.StartServerAndNewClient(rpcAddress, confFilePath)
-	if err != nil {
-		return nil, err
-	}
-	return &RPCProcessController{
-		client: client,
-	}, nil
-}
-
-// Status 获取所有进程的状态
-func (p *RPCProcessController) Status() ([]rpc.ProcessInfo, error) {
-	return p.client.GetAllProcessInfo()
-}
-
-// Stop 停止(不是删除)进程
-func (p *RPCProcessController) Stop(name string) error {
-	return p.client.StopProcess(name, true)
-}
-
-// Start 启动进程（只能操作已存在的进程）
-func (p *RPCProcessController) Start(name string) error {
-	return p.client.StartProcess(name, true)
-}
-
 // RefreshConf 重新生成配置文件
 func RefreshConf(processes []Process, procEnvs ...appdesc.Env) error {
 	conf, err := makeSupervisorConf(processes, procEnvs...)
@@ -225,13 +127,44 @@ func RefreshConf(processes []Process, procEnvs ...appdesc.Env) error {
 	return nil
 }
 
+// RPCProcessController ...
+type SupervisorRPCProcessController struct {
+	client *rpc.Client
+}
+
+// 创建 Supervisor RPC 类型的 ProcessController
+func newSupervisorRPCProcessController() (*SupervisorRPCProcessController, error) {
+	client := rpc.NewClient(rpcAddress)
+	return &SupervisorRPCProcessController{
+		client: client,
+	}, nil
+}
+
+// Status 获取所有进程的状态
+func (p *SupervisorRPCProcessController) Status() ([]rpc.ProcessInfo, error) {
+	return p.client.GetAllProcessInfo()
+}
+
+// Stop 停止(不是删除)进程
+func (p *SupervisorRPCProcessController) Stop(name string) error {
+	return p.client.StopProcess(name, true)
+}
+
+// Start 启动进程（只能操作已存在的进程）
+func (p *SupervisorRPCProcessController) Start(name string) error {
+	return p.client.StartProcess(name, true)
+}
+
 // Reload 更新和重启进程列表
-func (p *RPCProcessController) Reload() error {
+func (p *SupervisorRPCProcessController) Reload(processes []Process, procEnvs ...appdesc.Env) error {
+	if err := RefreshConf(processes, procEnvs...); err != nil {
+		return err
+	}
 	return p.client.Restart()
 }
 
 // StopAllProcesses 停止所有进程
-func (p *RPCProcessController) StopAllProcesses() error {
+func (p *SupervisorRPCProcessController) StopAllProcesses() error {
 	_, err := p.client.StopAllProcesses(true)
 	return err
 }
