+++ cnb-builder-shim/internal/devsandbox/reload.go
@@ -76,7 +76,7 @@ func (m HotReloadManager) Rebuild(reloadID string) error {
 	// 2. 构建时，supervisor 会因 terminated by SIGSEGV 发生预期之外进程退出，
 	// 然后本身的托管能力又会拉起进程，导致 /app/v3logs 目录权限在 cnb or root
 	// 间横跳，从而使构建时遇到 v3logs 文件夹权限错误而构建失败
-	processCtl, err := processesctl.NewProcessController(processesctl.RPC)
+	processCtl, err := processesctl.NewProcessController()
 	if err == nil {
 		_ = processCtl.StopAllProcesses()
 	}
